refactor: users API 수정
