--- supabase/functions/api/supabase.yaml
@@ -1,4 +1,0 @@
-functions:
-  api:
-    handler:
-      src: functions/api/index.ts
